+++ src/main/java/com/irontechspace/dynamicdq/annotations/ExecDuration.java
@@ -5,10 +5,10 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target({ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
 public @interface ExecDuration {
-    String param() default "";
+    String[] params() default {};
 }