+++ src/main/java/com/irontechspace/dynamicdq/aspects/AsyncConfiguration.java
@@ -1,0 +1,25 @@
+package com.irontechspace.dynamicdq.aspects;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.util.concurrent.Executor;
+
+@Configuration
+@EnableAsync
+public class AsyncConfiguration
+{
+    @Bean(name = "AsyncExecDuration")
+    public Executor asyncExecutor()
+    {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(3);
+        executor.setMaxPoolSize(3);
+        executor.setQueueCapacity(100);
+        executor.setThreadNamePrefix("ExecDuration-");
+        executor.initialize();
+        return executor;
+    }
+}