+++ src/main/java/com/irontechspace/dynamicdq/configurator/query/QueryConfigRowMapper.java
@@ -15,6 +15,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
+import static com.irontechspace.dynamicdq.exceptions.ExceptionUtils.logException;
 import static com.irontechspace.dynamicdq.utils.SqlUtils.inspect;
 
 @Log4j2
@@ -40,7 +41,8 @@ public class QueryConfigRowMapper implements RowMapper<QueryConfig> {
                 } else
                     bw.setPropertyValue(field, value);
             } catch (Exception e) {
-                e.printStackTrace();
+//                e.printStackTrace();
+                logException(log, e);
             }
         }
         return mappedObject;