+++ src/main/java/com/irontechspace/dynamicdq/executor/export/model/ExcelBorder.java
@@ -1,0 +1,13 @@
+package com.irontechspace.dynamicdq.executor.export.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ExcelBorder {
+    private Boolean top;
+    private Boolean right;
+    private Boolean bottom;
+    private Boolean left;
+}