+++ src/main/java/com/irontechspace/dynamicdq/executor/export/model/ExcelFont.java
@@ -1,0 +1,11 @@
+package com.irontechspace.dynamicdq.executor.export.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ExcelFont {
+    private Integer size;
+    private Boolean bold;
+}