+++ src/main/java/com/irontechspace/dynamicdq/executor/task/TaskService.java
@@ -2,7 +2,9 @@ package com.irontechspace.dynamicdq.executor.task;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.irontechspace.dynamicdq.executor.events.SystemEventsService;
+import com.irontechspace.dynamicdq.executor.export.ExportExcelService;
 import com.irontechspace.dynamicdq.executor.task.model.*;
 import com.irontechspace.dynamicdq.executor.query.QueryService;
 import com.irontechspace.dynamicdq.executor.save.SaveService;
@@ -18,6 +20,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
+
+import static com.irontechspace.dynamicdq.exceptions.ExceptionUtils.logException;
 
 @Log4j2
 @Service
@@ -35,15 +39,23 @@ public class TaskService {
     RabbitSender rabbitSender;
 
     @Autowired
+    SimpleTaskConfigs simpleTaskConfigs;
+
+    @Autowired
     TaskUtils taskUtils;
 
     @Autowired
     SystemEventsService systemEventsService;
 
+    @Autowired
+    ExportExcelService exportExcelService;
+
     public Object executeTask(Task task) {
 
-        Map<String, Object> outputData = new HashMap<>();
+//        Map<String, Object> outputData = new HashMap<>();
+        ObjectNode outputData = OBJECT_MAPPER.createObjectNode();
         for (int i = 0; i < task.getConfigs().size(); i++) {
+
 
             // Init config params
             TaskConfig config = task.getConfigs().get(i);
@@ -61,16 +73,18 @@ public class TaskService {
                 // Любой конфиг кроме выходного (branch пока исключен совсем)
                 if (config.getTypeExecutor() != TaskType.output) {
                     Object res = executeConfig(config.getTypeExecutor(), config.getConfigName(), task.getUserId(), task.getUserRoles(), body, pageable);
-                    if (config.getOutput() != null) outputData.put(config.getOutput(), res);
+                    if (config.getOutput() != null)
+                        taskUtils.setOutputData(config.getTypeExecutor(), config.getOutput().split("\\."), outputData, res);
                 }
 
                 // FINISH Event
                 if (events != null) executeEvent("finish", task.getUserId(), events.getFinish(), body);
 
                 // Выходной конфиг
-                if (config.getTypeExecutor() == TaskType.output) return body;
+                if (config.getTypeExecutor() == TaskType.output)
+                    return body;
             } catch (Exception e) {
-                e.printStackTrace();
+                logException(log, e);
                 if (events != null)
                     executeEvent("error", task.getUserId(), events.getError(), OBJECT_MAPPER.createObjectNode().put("error", e.getMessage()));
             }
@@ -107,14 +121,18 @@ public class TaskService {
             case save:
                 return saveService.saveData(configName, userId, userRoles, body);
             case queue: return rabbitSender.send(configName, body);
-            case equal: return taskUtils.compare(TaskType.equal, body);
-            case notEqual: return taskUtils.compare(TaskType.notEqual, body);
-            case greaterEqual: return taskUtils.compare(TaskType.greaterEqual, body);
-            case lessEqual: return taskUtils.compare(TaskType.lessEqual, body);
-            case greater: return taskUtils.compare(TaskType.greater, body);
-            case less: return taskUtils.compare(TaskType.less, body);
-            case log: return taskUtils.log(body);
+            case equal: return simpleTaskConfigs.compare(TaskType.equal, body);
+            case notEqual: return simpleTaskConfigs.compare(TaskType.notEqual, body);
+            case greaterEqual: return simpleTaskConfigs.compare(TaskType.greaterEqual, body);
+            case lessEqual: return simpleTaskConfigs.compare(TaskType.lessEqual, body);
+            case greater: return simpleTaskConfigs.compare(TaskType.greater, body);
+            case less: return simpleTaskConfigs.compare(TaskType.less, body);
+            case log: return simpleTaskConfigs.log(body);
             case event: return executeEvent(userId, body);
+            case createExcel: return exportExcelService.createExcel(body);
+            case addTableExcel: return exportExcelService.addTableExcel(configName, userId, userRoles, body);
+            case addRowExcel: return exportExcelService.addRowExcel(body);
+            case saveExcel: return exportExcelService.saveExcel(configName, userId, userRoles, body);
             default:
                 throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Ошибка запроса. Указан не существующий execute type");
         }