+++ src/main/java/com/irontechspace/dynamicdq/rabbit/RabbitReceiver.java
@@ -8,6 +8,8 @@ import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import static com.irontechspace.dynamicdq.exceptions.ExceptionUtils.logException;
+
 @Log4j2
 @Component
 public class RabbitReceiver {
@@ -30,7 +32,8 @@ public class RabbitReceiver {
             log.info("Received notification:\t{}", notification.toString());
         }
         catch (Exception e){
-            e.printStackTrace();
+//            e.printStackTrace();
+            logException(log, e);
         }
     }